refactor(topk): relocate top-k into flashinfer/top_k/ package
Move the CuTe-DSL top-k kernels and the top_k_varlen API out of the
shared flashinfer/cute_dsl/ tree into a dedicated flashinfer/top_k/
package, mirroring flashinfer/quantization/ (kernels under kernels/).
Addresses PR review comments about not accumulating feature-specific
code under flashinfer/cute_dsl/.

  flashinfer/cute_dsl/top_k/*   -> flashinfer/top_k/kernels/*
  flashinfer/topk_varlen.py     -> flashinfer/top_k/topk_varlen.py
  tests/topk/                   -> tests/top_k/   (name consistency)

Kernel sources are byte-identical (git mv only); only import paths and
the ruff exclude list in pyproject.toml change. The public API
flashinfer.top_k_varlen is unchanged (re-exported from __init__.py).

Verified: pre-commit clean; tests pass on sm_80/89/90/100/120
(115 passed on sm_100, exercising GVR + radix from the new path);
trace-registry inventory green.
