+++ flashinfer/__init__.py
@@ -232,7 +232,7 @@ from .topk import top_k as top_k
 from .topk import top_k_page_table_transform as top_k_page_table_transform
 from .topk import top_k_ragged_transform as top_k_ragged_transform
 from .topk import TopKTieBreak as TopKTieBreak
-from .topk_varlen import top_k_varlen as top_k_varlen
+from .top_k.topk_varlen import top_k_varlen as top_k_varlen
 from .sparse import BlockSparseAttentionWrapper as BlockSparseAttentionWrapper
 from .sparse import (
     VariableBlockSparseAttentionWrapper as VariableBlockSparseAttentionWrapper,
