+++ flashinfer/top_k/__init__.py
@@ -1,0 +1,21 @@
+# Copyright (c) 2025 by FlashInfer team.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#   http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+"""Top-K decode kernels for FlashInfer.
+
+The CuTe-DSL kernel source (GVR and radix Top-K for Blackwell sm_100+) lives in
+``flashinfer.top_k.kernels``.  The public ``top_k_varlen`` API is defined in
+``flashinfer.top_k.topk_varlen`` and re-exported from the top-level
+``flashinfer`` namespace.
+"""
